Create BST search
The base case will begin with null, or if the key is present at the root. After starting at the root, it will compare the searching element with the root. If it's less than the root, it will either call the left of right subtree. If the element that is being searched is found, the code will either return true or false.